% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailcall_eliminate.R
\name{tailcall_eliminate}
\alias{tailcall_eliminate}
\title{Eliminates tailcalls under special circumstances}
\usage{
tailcall_eliminate(fun)
}
\arguments{
\item{fun}{a function}
}
\value{
the same function, but with more efficient byte-code.
If the function was optimized.
}
\description{
\code{tailcall_eliminate} takes a function. If that functions has a \code{Recall} call as
the last operation, it will attempt to replace the recursion by a \code{GOTO} statement
in the byte-code representation. Beware, this can have unintended problems, if
the calls have side-effects.
}
